emotion_understanding/requirements.txt
# conda create -n emotion_reco python=3.8
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python
pyrealsense2
numpy
Pillow
scikit-learn
matplotlib
tqdm
seaborn